Fix a type alignment/casting issue. It turns out the ancillary data returned from IP_TTL is an int, not a uchar. who knew? (big endien test host found it)
git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@1001 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ owamp/endpoint.c
@@ -1807,6 +1807,13 @@ recvfromttl(
                  * defined.
                  *
                  * Gotta love standards...
+                 *
+                 * (Looking at opendarwin kernel sources leads me to
+                 * believe that IP_RECVTTL is a uchar and actual
+                 * documentation in the CMSG man page on Linux
+                 * tells me IP_TTL should be treated as an int. But,
+                 * I have never really found the 'definitive' standard
+                 * for this stuff. oh well...)
                  */
 #ifdef  IP_RECVTTL
                 if(cmdmsgptr->cmsg_level == IPPROTO_IP &&
@@ -1818,7 +1825,7 @@ recvfromttl(
 #endif
                 if(cmdmsgptr->cmsg_level == IPPROTO_IP &&
                         cmdmsgptr->cmsg_type == IP_TTL){
-                    *ttl = *(uint8_t *)CMSG_DATA(cmdmsgptr);
+                    *ttl = *(int *)CMSG_DATA(cmdmsgptr);
                     goto NEXTCMSG;
                 }
                 break;
